added improvements to release notes
